Prevent Skill crash on non-mark-ii device
This moves the import of a Mark II only function to a try block.
Whilst it should be expected on a real Mark II. If you are
developing on another machine, this prevents being able to run the
Mark II Skill and thus emulate the Mark II display as well.